requirements.txt
feapder==1.8.6
pymongo==4.5.0
bs4~=0.0.1
newspaper3k==0.2.8
langdetect==1.0.9
playwright==1.40.0
trafilatura==1.6.4
htmldate==1.6.1
courlan==0.9.5
unstructured==0.11.8
Flask==3.0.0